refactor: pages/Follower -> components로 이전
