+++ apps/web/app/follower/select-profile/page.ts
@@ -1,2 +1,2 @@
-import Page from "@/pages/Follower/SelectProfile/Page";
+import Page from "@/components/follower-select-profile/FollowerSelectProfilePage";
 export default Page;
